requirements.txt
asn1crypto==0.22.0
certifi==2017.7.27.1
cffi==1.10.0
chardet==3.0.4
cryptography==2.0.2
dj-database-url==0.4.2
dj-static==0.0.6
django==1.11.2
django-assets==0.12
django-ckeditor==5.2.2
django-filter==1.0.4
django-tinymce==2.6.0
django-wysiwyg==0.8.0
djangorestframework==3.6.3
gspread==0.6.2
gunicorn==19.7.1
httplib2==0.10.3
idna==2.5
Markdown==2.6.8
mysql-connector==2.1.4
oauth2client==4.1.2
olefile==0.44
Pillow==4.1.1
psycopg2==2.7.3
pyasn1==0.3.1
pyasn1-modules==0.0.10
pycparser==2.18
pyOpenSSL==17.2.0
pytz==2017.2
requests==2.18.3
rsa==3.4.2
simplejson==3.11.1
six==1.10.0
static3==0.7.0
urllib3==1.22
virtualenv==15.1.0
webassets==0.12.1
whitenoise==3.3.0
